Use only new EquivalentExpression
